requirements.txt
Django>=5.2.4
psycopg2-binary>=2.9.9
Pillow>=10.0.0 